+++ bon-macros/src/builder/item_struct.rs
@@ -1,15 +1,11 @@
-use super::builder_gen::input_struct::{StructInputCtx, StructInputParams};
+use super::builder_gen::input_struct::StructInputCtx;
 use super::builder_gen::MacroOutput;
 use crate::util::prelude::*;
 use quote::quote;
 
-pub(crate) fn generate(
-    params: StructInputParams,
-    orig_struct: syn::ItemStruct,
-) -> Result<TokenStream2> {
-    let ctx = StructInputCtx::new(params, orig_struct);
-
-    let adapted_struct = ctx.adapted_struct();
+pub(crate) fn generate(orig_struct: syn::ItemStruct) -> Result<TokenStream2> {
+    let struct_ident = orig_struct.ident.clone();
+    let ctx = StructInputCtx::new(orig_struct)?;
 
     let MacroOutput {
         mut start_func,
@@ -20,8 +16,6 @@ pub(crate) fn generate(
 
     let (generics_decl, generic_args, where_clause) = impl_generics.split_for_impl();
 
-    let struct_ident = &adapted_struct.ident;
-
     Ok(quote! {
         #[automatically_derived]
         impl #generics_decl #struct_ident #generic_args
@@ -31,6 +25,5 @@ pub(crate) fn generate(
         }
 
         #other_items
-        #adapted_struct
     })
 }